refactor(gemini): reorganize invoke tests by outcome
